trezor dir restructuring
