+++ app/assets/images/trezor/connect/connect-load.gif
