+++ app/assets/images/trezor/connect/connect-start.gif
