+++ app/components/wallet/send/WalletSendConfirmationDialog.scss
@@ -1,5 +1,5 @@
-@import '../../themes/mixins/loading-spinner';
-@import '../../themes/mixins/error-message';
+@import '../../../themes/mixins/loading-spinner';
+@import '../../../themes/mixins/error-message';
 
 .dialog {
   font-family: var(--font-medium);
@@ -101,5 +101,5 @@
 }
 
 .submitButtonSpinning {
-  @include loading-spinner("../../assets/images/spinner-light.svg");
+  @include loading-spinner("../../../assets/images/spinner-light.svg");
 }
