+++ app/components/wallet/send/WalletSendForm.scss
@@ -1,5 +1,5 @@
-@import '../../themes/mixins/loading-spinner';
-@import '../../themes/mixins/error-message';
+@import '../../../themes/mixins/loading-spinner';
+@import '../../../themes/mixins/error-message';
 
 .component {
   padding: 20px;
