+++ app/components/wallet/trezorConnect/AboutDialog.js
@@ -16,8 +16,8 @@ import HelpLinkBlock from './common/HelpLinkBlock';
 import ErrorBlock from './common/ErrorBlock';
 
 import externalLinkSVG from '../../../assets/images/link-external.inline.svg';
-import aboutPrerequisiteIconSVG from '../../../assets/images/trezor/about-prerequisite-header-icon.inline.svg';
-import aboutPrerequisiteTrezorSVG from '../../../assets/images/trezor/about-trezor.inline.svg';
+import aboutPrerequisiteIconSVG from '../../../assets/images/trezor/connect/about-prerequisite-header-icon.inline.svg';
+import aboutPrerequisiteTrezorSVG from '../../../assets/images/trezor/connect/about-trezor.inline.svg';
 
 import type { ProgressInfo } from '../../../stores/ada/TrezorConnectStore';
