+++ app/components/wallet/trezorConnect/common/DialogCommonStyle.scss
@@ -1,4 +1,4 @@
-@import '../../../themes/mixins/loading-spinner.scss';
+@import '../../../../themes/mixins/loading-spinner.scss';
 
 $opacityCommon: 0.7;
